Introduce actions-util.ts
