+++ lib/api-client.js
@@ -14,6 +14,7 @@ const core = __importStar(require("@actions/core"));
 const github = __importStar(require("@actions/github"));
 const console_log_level_1 = __importDefault(require("console-log-level"));
 const path = __importStar(require("path"));
+const actions_util_1 = require("./actions-util");
 const util_1 = require("./util");
 exports.getApiClient = function (githubAuth, githubUrl, allowLocalRun = false) {
     if (util_1.isLocalRun() && !allowLocalRun) {
@@ -41,7 +42,7 @@ function getApiUrl(githubUrl) {
 // Once all code has been coverted this function should be removed or made canonical
 // and called only from the action entrypoints.
 function getActionsApiClient(allowLocalRun = false) {
-    return exports.getApiClient(core.getInput("token"), util_1.getRequiredEnvParam("GITHUB_SERVER_URL"), allowLocalRun);
+    return exports.getApiClient(core.getInput("token"), actions_util_1.getRequiredEnvParam("GITHUB_SERVER_URL"), allowLocalRun);
 }
 exports.getActionsApiClient = getActionsApiClient;
 //# sourceMappingURL=api-client.js.map
