+++ lib/codeql.js
@@ -19,6 +19,7 @@ const semver = __importStar(require("semver"));
 const stream = __importStar(require("stream"));
 const globalutil = __importStar(require("util"));
 const v4_1 = __importDefault(require("uuid/v4"));
+const actions_util_1 = require("./actions-util");
 const api = __importStar(require("./api-client"));
 const defaults = __importStar(require("./defaults.json")); // Referenced from codeql-action-sync-tool!
 const error_matcher_1 = require("./error-matcher");
@@ -39,7 +40,7 @@ function getCodeQLActionRepository(mode) {
     // Actions do not know their own repository name,
     // so we currently use this hack to find the name based on where our files are.
     // This can be removed once the change to the runner in https://github.com/actions/runner/pull/585 is deployed.
-    const runnerTemp = util.getRequiredEnvParam("RUNNER_TEMP");
+    const runnerTemp = actions_util_1.getRequiredEnvParam("RUNNER_TEMP");
     const actionsDirectory = path.join(path.dirname(runnerTemp), "_actions");
     const relativeScriptPath = path.relative(actionsDirectory, __filename);
     // This handles the case where the Action does not come from an Action repository,
