+++ lib/init-action.js
@@ -8,12 +8,12 @@ var __importStar = (this && this.__importStar) || function (mod) {
 };
 Object.defineProperty(exports, "__esModule", { value: true });
 const core = __importStar(require("@actions/core"));
+const actionsUtil = __importStar(require("./actions-util"));
 const init_1 = require("./init");
 const logging_1 = require("./logging");
 const repository_1 = require("./repository");
-const util = __importStar(require("./util"));
 async function sendSuccessStatusReport(startedAt, config) {
-    const statusReportBase = await util.createStatusReportBase("init", "success", startedAt);
+    const statusReportBase = await actionsUtil.createStatusReportBase("init", "success", startedAt);
     const languages = config.languages.join(",");
     const workflowLanguages = core.getInput("languages", { required: false });
     const paths = (config.originalUserInput.paths || []).join(",");
@@ -33,7 +33,7 @@ async function sendSuccessStatusReport(startedAt, config) {
         disable_default_queries: disableDefaultQueries,
         queries,
     };
-    await util.sendStatusReport(statusReport);
+    await actionsUtil.sendStatusReport(statusReport);
 }
 async function run() {
     const startedAt = new Date();
@@ -41,17 +41,17 @@ async function run() {
     let config;
     let codeql;
     try {
-        util.prepareLocalRunEnvironment();
-        if (!(await util.sendStatusReport(await util.createStatusReportBase("init", "starting", startedAt), true))) {
+        actionsUtil.prepareLocalRunEnvironment();
+        if (!(await actionsUtil.sendStatusReport(await actionsUtil.createStatusReportBase("init", "starting", startedAt), true))) {
             return;
         }
-        codeql = await init_1.initCodeQL(core.getInput("tools"), core.getInput("token"), util.getRequiredEnvParam("GITHUB_SERVER_URL"), util.getRequiredEnvParam("RUNNER_TEMP"), util.getRequiredEnvParam("RUNNER_TOOL_CACHE"), "actions", logger);
-        config = await init_1.initConfig(core.getInput("languages"), core.getInput("queries"), core.getInput("config-file"), repository_1.parseRepositoryNwo(util.getRequiredEnvParam("GITHUB_REPOSITORY")), util.getRequiredEnvParam("RUNNER_TEMP"), util.getRequiredEnvParam("RUNNER_TOOL_CACHE"), codeql, util.getRequiredEnvParam("GITHUB_WORKSPACE"), core.getInput("token"), util.getRequiredEnvParam("GITHUB_SERVER_URL"), logger);
+        codeql = await init_1.initCodeQL(core.getInput("tools"), core.getInput("token"), actionsUtil.getRequiredEnvParam("GITHUB_SERVER_URL"), actionsUtil.getRequiredEnvParam("RUNNER_TEMP"), actionsUtil.getRequiredEnvParam("RUNNER_TOOL_CACHE"), "actions", logger);
+        config = await init_1.initConfig(core.getInput("languages"), core.getInput("queries"), core.getInput("config-file"), repository_1.parseRepositoryNwo(actionsUtil.getRequiredEnvParam("GITHUB_REPOSITORY")), actionsUtil.getRequiredEnvParam("RUNNER_TEMP"), actionsUtil.getRequiredEnvParam("RUNNER_TOOL_CACHE"), codeql, actionsUtil.getRequiredEnvParam("GITHUB_WORKSPACE"), core.getInput("token"), actionsUtil.getRequiredEnvParam("GITHUB_SERVER_URL"), logger);
     }
     catch (e) {
         core.setFailed(e.message);
         console.log(e);
-        await util.sendStatusReport(await util.createStatusReportBase("init", "aborted", startedAt, e.message));
+        await actionsUtil.sendStatusReport(await actionsUtil.createStatusReportBase("init", "aborted", startedAt, e.message));
         return;
     }
     try {
@@ -75,7 +75,7 @@ async function run() {
     catch (error) {
         core.setFailed(error.message);
         console.log(error);
-        await util.sendStatusReport(await util.createStatusReportBase("init", "failure", startedAt, error.message, error.stack));
+        await actionsUtil.sendStatusReport(await actionsUtil.createStatusReportBase("init", "failure", startedAt, error.message, error.stack));
         return;
     }
     await sendSuccessStatusReport(startedAt, config);
