+++ lib/util.test.js
@@ -62,10 +62,6 @@ ava_1.default("getThreadsFlag() should return the correct --threads flag", (t) =
 ava_1.default("getThreadsFlag() throws if the threads input is not an integer", (t) => {
     t.throws(() => util.getThreadsFlag("hello!", logging_1.getRunnerLogger(true)));
 });
-ava_1.default("getRef() throws on the empty string", (t) => {
-    process.env["GITHUB_REF"] = "";
-    t.throws(util.getRef);
-});
 ava_1.default("isLocalRun() runs correctly", (t) => {
     const origLocalRun = process.env.CODEQL_LOCAL_RUN;
     process.env.CODEQL_LOCAL_RUN = "";
@@ -78,23 +74,6 @@ ava_1.default("isLocalRun() runs correctly", (t) => {
     t.assert(util.isLocalRun());
     process.env.CODEQL_LOCAL_RUN = "hucairz";
     t.assert(util.isLocalRun());
-    process.env.CODEQL_LOCAL_RUN = origLocalRun;
-});
-ava_1.default("prepareEnvironment() when a local run", (t) => {
-    const origLocalRun = process.env.CODEQL_LOCAL_RUN;
-    process.env.CODEQL_LOCAL_RUN = "false";
-    process.env.GITHUB_JOB = "YYY";
-    util.prepareLocalRunEnvironment();
-    // unchanged
-    t.deepEqual(process.env.GITHUB_JOB, "YYY");
-    process.env.CODEQL_LOCAL_RUN = "true";
-    util.prepareLocalRunEnvironment();
-    // unchanged
-    t.deepEqual(process.env.GITHUB_JOB, "YYY");
-    process.env.GITHUB_JOB = "";
-    util.prepareLocalRunEnvironment();
-    // updated
-    t.deepEqual(process.env.GITHUB_JOB, "UNKNOWN-JOB");
     process.env.CODEQL_LOCAL_RUN = origLocalRun;
 });
 ava_1.default("getExtraOptionsEnvParam() succeeds on valid JSON with invalid options (for now)", (t) => {
