+++ src/upload-sarif-action.ts
@@ -1,19 +1,19 @@
 import * as core from "@actions/core";
 
+import * as actionsUtil from "./actions-util";
 import { getActionsLogger } from "./logging";
 import { parseRepositoryNwo } from "./repository";
 import * as upload_lib from "./upload-lib";
-import * as util from "./util";
 
 interface UploadSarifStatusReport
-  extends util.StatusReportBase,
+  extends actionsUtil.StatusReportBase,
     upload_lib.UploadStatusReport {}
 
 async function sendSuccessStatusReport(
   startedAt: Date,
   uploadStats: upload_lib.UploadStatusReport
 ) {
-  const statusReportBase = await util.createStatusReportBase(
+  const statusReportBase = await actionsUtil.createStatusReportBase(
     "upload-sarif",
     "success",
     startedAt
@@ -22,14 +22,18 @@ async function sendSuccessStatusReport(
     ...statusReportBase,
     ...uploadStats,
   };
-  await util.sendStatusReport(statusReport);
+  await actionsUtil.sendStatusReport(statusReport);
 }
 
 async function run() {
   const startedAt = new Date();
   if (
-    !(await util.sendStatusReport(
-      await util.createStatusReportBase("upload-sarif", "starting", startedAt),
+    !(await actionsUtil.sendStatusReport(
+      await actionsUtil.createStatusReportBase(
+        "upload-sarif",
+        "starting",
+        startedAt
+      ),
       true
     ))
   ) {
@@ -39,16 +43,16 @@ async function run() {
   try {
     const uploadStats = await upload_lib.upload(
       core.getInput("sarif_file"),
-      parseRepositoryNwo(util.getRequiredEnvParam("GITHUB_REPOSITORY")),
-      await util.getCommitOid(),
-      util.getRef(),
-      await util.getAnalysisKey(),
-      util.getRequiredEnvParam("GITHUB_WORKFLOW"),
-      util.getWorkflowRunID(),
+      parseRepositoryNwo(actionsUtil.getRequiredEnvParam("GITHUB_REPOSITORY")),
+      await actionsUtil.getCommitOid(),
+      actionsUtil.getRef(),
+      await actionsUtil.getAnalysisKey(),
+      actionsUtil.getRequiredEnvParam("GITHUB_WORKFLOW"),
+      actionsUtil.getWorkflowRunID(),
       core.getInput("checkout_path"),
       core.getInput("matrix"),
       core.getInput("token"),
-      util.getRequiredEnvParam("GITHUB_SERVER_URL"),
+      actionsUtil.getRequiredEnvParam("GITHUB_SERVER_URL"),
       "actions",
       getActionsLogger()
     );
@@ -56,8 +60,8 @@ async function run() {
   } catch (error) {
     core.setFailed(error.message);
     console.log(error);
-    await util.sendStatusReport(
-      await util.createStatusReportBase(
+    await actionsUtil.sendStatusReport(
+      await actionsUtil.createStatusReportBase(
         "upload-sarif",
         "failure",
         startedAt,
